+++ tests/mock/test_functions.py
@@ -4,8 +4,6 @@
 import datetime
 import math
 
-import pytest
-
 from snowflake.snowpark import DataFrame, Row
 from snowflake.snowpark._internal.type_utils import ColumnOrName
 from snowflake.snowpark.column import Column, _to_col_if_str
@@ -28,7 +26,6 @@ from snowflake.snowpark.mock._functions import patch
 from snowflake.snowpark.mock._snowflake_data_type import ColumnEmulator
 
 
-@pytest.mark.localtest
 def test_col(session):
     origin_df: DataFrame = session.create_dataframe(
         [
@@ -43,7 +40,6 @@ def test_col(session):
     assert origin_df.select(col("o")).collect() == [Row(True), Row(False), Row(None)]
 
 
-@pytest.mark.localtest
 def test_max(session):
     origin_df: DataFrame = session.create_dataframe(
         [
@@ -64,7 +60,6 @@ def test_max(session):
     assert math.isnan(origin_df.select(max("s").as_("g")).collect()[0][0])
 
 
-@pytest.mark.localtest
 def test_min(session):
     origin_df: DataFrame = session.create_dataframe(
         [
@@ -86,7 +81,6 @@ def test_min(session):
     assert math.isnan(origin_df.select(min("s").as_("g")).collect()[0][0])
 
 
-@pytest.mark.localtest
 def test_to_date(session):
     origin_df: DataFrame = session.create_dataframe(
         ["2013-05-17", "31536000000000"],
@@ -99,7 +93,6 @@ def test_to_date(session):
     ]
 
 
-@pytest.mark.localtest
 def test_contains(session):
     origin_df: DataFrame = session.create_dataframe(
         [
@@ -138,7 +131,6 @@ def test_contains(session):
     ]
 
 
-@pytest.mark.localtest
 def test_abs(session):
     origin_df: DataFrame = session.create_dataframe(
         [
@@ -151,7 +143,6 @@ def test_abs(session):
     assert origin_df.select(abs(col("m"))).collect() == [Row(1), Row(1), Row(2)]
 
 
-@pytest.mark.localtest
 def test_asc_and_desc(session):
     origin_df: DataFrame = session.create_dataframe(
         [
@@ -170,7 +161,6 @@ def test_asc_and_desc(session):
     assert origin_df.sort(desc(col("v"))).collect() == expected
 
 
-@pytest.mark.localtest
 def test_count(session):
     origin_df: DataFrame = session.create_dataframe(
         [
@@ -186,7 +176,6 @@ def test_count(session):
     assert origin_df.select(count("v")).collect() == [Row(6)]
 
 
-@pytest.mark.localtest
 def test_is_null(session):
     origin_df: DataFrame = session.create_dataframe(
         [
@@ -207,7 +196,6 @@ def test_is_null(session):
     ]
 
 
-@pytest.mark.localtest
 def test_take_first(session):
     origin_df: DataFrame = session.create_dataframe(
         [
@@ -249,7 +237,6 @@ def test_take_first(session):
     assert math.isnan(res[4][0]) and res[4][1] == 200 and res[4][2] is None
 
 
-@pytest.mark.localtest
 def test_show(session):
     origin_df: DataFrame = session.create_dataframe(
         [